+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Fixed
 
-- Fixed issue finding nested assocated data caused by [3.5.0] and added additional tests
+- Fixed issues with overriding associated conditions [3.5.0] and improved testing
 
 ## [3.6.0] - 2020-08-31